Change StateMachineRet types
 - add a Response Type
 - add a Request Type (Future use - for a prompt callback)
 - change all programmers to use Response instead of bool
   - you can send any response, but the receiver won't let you add types
   you aren't expecting so you can still make errors writing resolve

TODO: write a asker() and use it to call a prompt callback supplied by
the selector, pushing the prompt code out of the state machine. It's a
simple promise based interface that has no idea what is responding.

This adds a new state where validator can request values and receive a
response before continuing.